dairy2imp/requirements.txt
Flask
mysql-connector-python
Flask-Login
